changed type of min max

+++ contracts/NFTSale.sol
@@ -116,8 +116,8 @@ contract NFTSale is Ownable {
         uint256 diceRollCost,
         uint8 houseEdge,
         uint256 revealCompensation,
-        uint8 min,
-        uint8 max
+        uint16 min,
+        uint16 max
     ) external payable {
         if (bytes(coupon).length != 0) {
             //Check if the coupon is valid and if not then revert
